WP AUTO UPDATE 2026-03-02
- Regenerated/updated all files under wp-content as part of automatic
  WordPress update performed on 2026‑03‑02.

Update docker-compose.yml

- Update docker compose to add the var for wordpress debug

+++ wp-content/plugins/suretriggers/src/Integrations/memberpress/actions/remove-membership.php
@@ -76,9 +76,17 @@ class RemoveMembership extends AutomateAction {
 	 * @return array|bool
 	 */
 	public function _action_listener( $user_id, $automation_id, $fields, $selected_options ) {
-
+		if ( ! class_exists( 'MeprSubscription' ) || ! class_exists( 'MeprHooks' ) || ! class_exists( 'MeprTransaction' ) ) {
+			return [
+				'status'  => 'error',
+				'message' => 'MemberPress classes not found.',
+			];
+		}
 		if ( ! $user_id ) {
-			throw new Exception( 'User not found with this email address.' );
+			return [
+				'status'  => 'error',
+				'message' => 'User not found with this email address.',
+			];
 		}
 
 		if ( is_array( $selected_options['memberpressproduct'] ) ) {
@@ -87,7 +95,13 @@ class RemoveMembership extends AutomateAction {
 			$membership_id = $selected_options['memberpressproduct'];
 		}
 		$user_obj = get_user_by( 'id', $user_id );
-		$table    = MeprSubscription::account_subscr_table(
+		if ( ! $user_obj instanceof \WP_User ) {
+			return [
+				'status'  => 'error',
+				'message' => 'User not found.',
+			];
+		}
+		$table = MeprSubscription::account_subscr_table(
 			'created_at',
 			'DESC',
 			'',
@@ -107,24 +121,26 @@ class RemoveMembership extends AutomateAction {
 		);
 
 		if ( 0 === $table['count'] ) {
-			$this->set_error(
-				[
-					'msg' => __( 'Empty subscription table ', 'suretriggers' ),
-				]
-			);
-			return false;
+			return [
+				'status'  => 'error',
+				'message' => __( 'Empty subscription table ', 'suretriggers' ), 
+				
+			];
 		}
 
 		foreach ( $table['results'] as $row ) {
 			if ( $row->product_id === $membership_id || '-1' === $membership_id ) {
+				$sub = null;
 				if ( 'subscription' === $row->sub_type ) {
 					$sub = new MeprSubscription( $row->id );
 				} elseif ( 'transaction' === $row->sub_type ) {
 					$sub = new MeprTransaction( $row->id );
 				}
-				$sub->destroy();
-				$member = $sub->user();
-				$member->update_member_data();
+				if ( $sub ) {
+					$sub->destroy();
+					$member = $sub->user();
+					$member->update_member_data();
+				}
 			}
 		}
